pyproject.toml
[project]
name = "photoshare"
version = "0.1.0"
description = ""
authors = [
    {name = "gozilla",email = "wpdr5r@gmail.com"}
]
package-mode = false
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "uploadview (>=0.2.7,<0.3.0)",
    "utils (>=1.0.2,<2.0.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "django-qrcode (>=0.3,<0.4)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"